% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ags.R
\name{ags}
\alias{ags}
\title{Points assigned for saturated fatty acids}
\usage{
ags(x)
}
\arguments{
\item{x}{The content in saturated fatty acids (in g/100g).}
}
\value{
The points assigned for saturated fatty acids
}
\description{
Points assigned for saturated fatty acids
}
\examples{
ags(5)

}
