% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compo_N.R
\name{Compo_N}
\alias{Compo_N}
\title{Points assigned for unfavourable elements of the nutritional score (version before 2023)}
\usage{
Compo_N(x)
}
\arguments{
\item{x}{The data frame with for unfavourable elements of the nutritional score.}
}
\value{
The points assigned for unfavourable elements of the nutritional score
}
\description{
Points assigned for unfavourable elements of the nutritional score (version before 2023)
}
\examples{
\dontrun{
Compo_N(data)
}

}
