% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compo_N2023.R
\name{Compo_N2023}
\alias{Compo_N2023}
\title{Points assigned for unfavourable elements of the nutritional score (version after 2023)}
\usage{
Compo_N2023(x)
}
\arguments{
\item{x}{The data frame with for unfavourable elements of the nutritional score.}
}
\value{
The points assigned for unfavourable elements of the nutritional score
}
\description{
Points assigned for unfavourable elements of the nutritional score (version after 2023)
}
\examples{
\dontrun{
Compo_N2023(data)
}

}
