% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compo_P.R
\name{Compo_P}
\alias{Compo_P}
\title{Points assigned for favourable elements of the nutritional score (version before 2023)}
\usage{
Compo_P(x)
}
\arguments{
\item{x}{The data frame with for favourable elements of the nutritional score.}
}
\value{
The points assigned for favourable elements of the nutritional score
}
\description{
Points assigned for favourable elements of the nutritional score (version before 2023)
}
\examples{
\dontrun{
Compo_P(data)
}

}
