% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energie.R
\name{energie}
\alias{energie}
\title{Points assigned for energy density}
\usage{
energie(x)
}
\arguments{
\item{x}{The energy density (in kJ/100g).}
}
\value{
The points assigned for energy density
}
\description{
Points assigned for energy density
}
\examples{
energie(1000)

}
