% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etiquette.R
\name{etiquette}
\alias{etiquette}
\title{Edit dataframe with the mandatory nutritional declaration}
\usage{
etiquette(
  Energie = NULL,
  AGS = NULL,
  Sucres = NULL,
  Sel = NULL,
  Proteines = NULL,
  Fibres = NULL,
  FetL = NULL,
  Red_meat_product = NULL
)
}
\arguments{
\item{Energie}{Energy (in kJ/100g)}

\item{AGS}{Saturated fatty acids (in g/100g)}

\item{Sucres}{Sugar content  (in g/100g)}

\item{Sel}{Salt content  (in g/100g)}

\item{Proteines}{Protein content  (in g/100g)}

\item{Fibres}{Fiber content  (in g/100g)}

\item{FetL}{FL percent}

\item{Red_meat_product}{Boolean value if more 50 percent of red meat}
}
\value{
The nutritional score
}
\description{
This function edits dataframe with the mandatory nutritional declaration.
}
\examples{
\dontrun{
score_nut(data)
}

}
