% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibre.R
\name{fibre}
\alias{fibre}
\title{Points assigned for fiber}
\usage{
fibre(x)
}
\arguments{
\item{x}{The content in fiber (in g/100g).}
}
\value{
The points assigned for fiber
}
\description{
Points assigned for fiber
}
\examples{
fibre(2)

}
