% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FL.R
\name{FL}
\alias{FL}
\title{Points assigned for fruits and vegetables}
\usage{
FL(x)
}
\arguments{
\item{x}{The content in fruits and vegetables (in percent).}
}
\value{
The points assigned for fruits and vegetables
}
\description{
Points assigned for fruits and vegetables
}
\examples{
FL(50)

}
