% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutriscore_Comp.R
\name{nutriscore_Comp}
\alias{nutriscore_Comp}
\title{Return results of NutriScore (versions before and after 2023)}
\usage{
nutriscore_Comp(x)
}
\arguments{
\item{x}{The dataframe with the mandatory nutritional declaration
#'}
}
\value{
The results of NutriScore (versions before and after 2023)
}
\description{
This function returns the NutriScore calculate with algorithm before and after 2023 modification.
}
\examples{
\dontrun{
nutriscore_Comp(data)
}

}
