% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteine.R
\name{proteine}
\alias{proteine}
\title{Points assigned for protein (version before 2023)}
\usage{
proteine(x)
}
\arguments{
\item{x}{The content in protein (in g/100g).}
}
\value{
The points assigned for protein
}
\description{
Points assigned for protein (version before 2023)
}
\examples{
proteine(10)

}
