% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteine2023.R
\name{proteine2023}
\alias{proteine2023}
\title{Points assigned for protein (version after 2023)}
\usage{
proteine2023(x)
}
\arguments{
\item{x}{The content in protein (in g/100g).}
}
\value{
The points assigned for protein
}
\description{
Points assigned for protein (version after 2023)
}
\examples{
proteine2023(10)

}
