% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_nut.R
\name{score_nut}
\alias{score_nut}
\title{Return results of the nutritional score (versions before 2023)}
\usage{
score_nut(x)
}
\arguments{
\item{x}{The dataframe
#'}
}
\value{
The nutritional score
}
\description{
This function returns the nutritional score.
}
\examples{
\dontrun{
score_nut(data)
}

}
