% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_nut2023.R
\name{score_nut2023}
\alias{score_nut2023}
\title{Return results of the nutritional score (versions after 2023)}
\usage{
score_nut2023(x)
}
\arguments{
\item{x}{The dataframe
#'}
}
\value{
The nutritional score
}
\description{
This function returns the nutritional score.
}
\examples{
\dontrun{
score_nut2023(data)
}

}
