% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sel.R
\name{sel}
\alias{sel}
\title{Points assigned for salt (version before 2023)}
\usage{
sel(x)
}
\arguments{
\item{x}{The content in salt (in g/100g).}
}
\value{
The points assigned for salt
}
\description{
Points assigned for salt (version before 2023)
}
\examples{
sel(500)

}
