% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sel2023.R
\name{sel2023}
\alias{sel2023}
\title{Points assigned for salt (version after 2023)}
\usage{
sel2023(x)
}
\arguments{
\item{x}{The content in salt (in g/100g).}
}
\value{
The points assigned for salt
}
\description{
Points assigned for salt (version after 2023)
}
\examples{
sel2023(500)

}
