% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sucre.R
\name{sucre}
\alias{sucre}
\title{Points assigned for sugars (version before 2023)}
\usage{
sucre(x)
}
\arguments{
\item{x}{The content in sugars (in g/100g).}
}
\value{
The points assigned for sugars
}
\description{
Points assigned for sugars (version before 2023)
}
\examples{
sucre(5)

}
