% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sucre2023.R
\name{sucre2023}
\alias{sucre2023}
\title{Points assigned for sugars (version after 2023)}
\usage{
sucre2023(x)
}
\arguments{
\item{x}{The content in sugars (in g/100g).}
}
\value{
The points assigned for sugars
}
\description{
Points assigned for sugars (version after 2023)
}
\examples{
sucre2023(5)

}
